Done extraction

+++ src/main/java/com/ckcest/ebs/vici/base/FocusData.java
@@ -22,6 +22,6 @@ public class FocusData {
 	//每个focus的出现次数
 	public static Map<String,Integer> focus2Num = new HashMap<String,Integer>();
 	//每个focus的层次信息(累加的层次信息，最终要除以focus次数来确定平均层次信息)
-	public static Map<String,Integer> focus2Hie = new HashMap<String,Integer>();
+	public static Map<String,Double> focus2Hie = new HashMap<String,Double>();
 	
 }
